+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.10] - DEV
 ### Added
+- Add a shortcuts window.
 - Add an option to toggle the dark theme.
 
 ### Changed